+++ src/package.ts
@@ -19,6 +19,7 @@ export type PackageJson = {
   scripts: Record<string, string>;
   files?: string[];
   pinnedDependencies?: string[];
+  repository?: string;
 } & AnyJson;
 
 export type ChangedPackageVersions = Array<{
@@ -31,7 +32,7 @@ export type NpmPackage = {
   name: string;
   version: string;
   versions: string[];
-  'dist-tags': string[];
+  'dist-tags': Record<string, string>;
 } & AnyJson;
 
 export interface VersionValidation {
@@ -171,7 +172,7 @@ export class Package extends AsyncOptionalCreatable {
       name: this.name,
       version: this.packageJson.version,
       versions: [],
-      'dist-tags': [],
+      'dist-tags': {},
     };
   }
 }