+++ test/repository.test.ts
@@ -24,7 +24,8 @@ describe('SinglePackageRepo', () => {
   let execStub: sinon.SinonStub;
 
   beforeEach(async () => {
-    uxStub = (stubInterface<UX>($$.SANDBOX, {}) as unknown) as UX;
+    // eslint-disable-next-line prettier/prettier
+    uxStub = stubInterface<UX>($$.SANDBOX, {}) as unknown as UX;
   });
 
   describe('isReleasable', () => {
@@ -383,7 +384,8 @@ describe('LernaRepo', () => {
   let revertAllChangesStub: sinon.SinonStub;
 
   beforeEach(async () => {
-    uxStub = (stubInterface<UX>($$.SANDBOX, {}) as unknown) as UX;
+    // eslint-disable-next-line prettier/prettier
+    uxStub = stubInterface<UX>($$.SANDBOX, {}) as unknown as UX;
     // if this stub doesn't exist, the test will revert all of your unstaged changes
     stubMethod($$.SANDBOX, LernaRepo.prototype, 'revertUnstagedChanges').returns(null);
     revertAllChangesStub = stubMethod($$.SANDBOX, LernaRepo.prototype, 'revertAllChanges').returns(null);